further changes to the system, most notably there's now an on delete cascade for parent child tasks - if you delete the parent the children are deleted too

+++ resources/views/tasks/index.blade.php
@@ -7,6 +7,7 @@
     <div class="panel-body">
         <!-- Display Validation Errors -->
         @include('errors.errors')
+        @include('messages.flash')
 
         <table class="table">
             <thead>
@@ -19,24 +20,41 @@
             </thead>
             <tbody>
         @foreach ($tasks as $task)
-            <tr>
-                <td>
-                    {{ $task->id }}
-                </td>
-                <td>
-                    {{ $task->name }}
-                </td>
-                <td>
-                    {{ Form::open([
-                        'route' => ['tasks.destroy', $task->id],
-                        'method' => 'DELETE'
-                    ]) }}
-                    {{ Form::submit('delete', ['class' => 'btn btn-danger']) }}
-                    {{ Form::close() }}
-                </td>
-                <td>
-                    <a href="{{ route('tasks.edit', $task->id) }}" class="btn btn-primary">Edit Task</a>
-                </td>
-            </tr>
+            <!-- Show only the parent tasks -->
+            @if($task->task_id === null)
+                <tr>
+                    <td>
+                        {{ $task->id }}
+                    </td>
+                    <td>
+                        <a href="{{ route('tasks.show', $task->id) }}">
+                            {{ $task->name }}
+                        </a>
+
+                        <!-- Show child tasks, if any -->
+                        <ul>
+                        @foreach($task->children as $child)
+                        <li>
+                            <a href="{{ route('tasks.show', $child->id) }}">
+                                {{ $child->name }}
+                            </a>
+                        </li>
+                        @endforeach
+                        </ul>
+
+                    </td>
+                    <td>
+                        {{ Form::open([
+                            'route' => ['tasks.destroy', $task->id],
+                            'method' => 'DELETE'
+                        ]) }}
+                        {{ Form::submit('delete', ['class' => 'btn btn-danger']) }}
+                        {{ Form::close() }}
+                    </td>
+                    <td>
+                        <a href="{{ route('tasks.edit', $task->id) }}" class="btn btn-primary">Edit Task</a>
+                    </td>
+                </tr>
+            @endif
         @endforeach
     </div>
